+++ numpy/lib/index_tricks.pyi
@@ -44,6 +44,11 @@ from numpy.typing import (
     _ShapeLike,
 )
 
+from numpy.core.multiarray import (
+    unravel_index as unravel_index,
+    ravel_multi_index as ravel_multi_index,
+)
+
 if sys.version_info >= (3, 8):
     from typing import Literal, SupportsIndex
 else:
@@ -56,34 +61,6 @@ _TupType = TypeVar("_TupType", bound=Tuple[Any, ...])
 _ArrayType = TypeVar("_ArrayType", bound=ndarray[Any, Any])
 
 __all__: List[str]
-
-@overload
-def unravel_index(  # type: ignore[misc]
-    indices: Union[int, integer[Any]],
-    shape: _ShapeLike,
-    order: _OrderCF = ...
-) -> Tuple[intp, ...]: ...
-@overload
-def unravel_index(
-    indices: _ArrayLikeInt,
-    shape: _ShapeLike,
-    order: _OrderCF = ...
-) -> Tuple[NDArray[intp], ...]: ...
-
-@overload
-def ravel_multi_index(  # type: ignore[misc]
-    multi_index: Sequence[Union[int, integer[Any]]],
-    dims: _ShapeLike,
-    mode: Union[_ModeKind, Tuple[_ModeKind, ...]] = ...,
-    order: _OrderCF = ...
-) -> intp: ...
-@overload
-def ravel_multi_index(
-    multi_index: Sequence[_ArrayLikeInt],
-    dims: _ShapeLike,
-    mode: Union[_ModeKind, Tuple[_ModeKind, ...]] = ...,
-    order: _OrderCF = ...
-) -> NDArray[intp]: ...
 
 @overload
 def ix_(*args: _NestedSequence[_SupportsDType[_DType]]) -> Tuple[ndarray[Any, _DType], ...]: ...
